refactor: embed find-comment
